requirements.txt
nicegui
mysql-connector-python
bcrypt
python-dotenv
